# swagger.yaml
openapi: 3.0.0
info:
  title: School Medical Management System API
  version: 1.0.0
  description: API documentation for School Medical Management System
servers:
  - url: http://localhost:3000/api
paths:
...
  /auth/google/callback:
    get:
      tags:
        - Authentication
      summary: Handle Google OAuth callback
      description: Processes the Google OAuth callback and returns user information
      operationId: handleGoogleCallback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful authentication with Google
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginRequest'
  /auth/login:
    post:
      summary: Login with JWT
      tags:
        - Authentication
      description: Authenticates a user with username and password, returns JWT tokens
      operationId: loginWithJwt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123; HttpOnly; SameSite=Strict; Expires=...
              description: Sets refreshToken cookie
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      description: Registers a new user with local authentication
      operationId: registerWithJwt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User registered successfully
...
  /auth/forgot-password:
    post:
      summary: Forgot password
      tags:
        - Authentication
      description: Sends an OTP to the user's email for password reset
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
              required:
                - email
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OTP sent to your email successfully
  /auth/verify-otp:
    post:
      summary: Verify OTP
      tags:
        - Authentication
      description: Verifies the OTP sent to the user's email
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                token:
                  type: string
                  example: '123456'
              required:
                - email
                - token
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: OTP verified successfully
                  data:
                    type: object
                    properties:
                      resetToken:
                        type: string
                        example: 'a1b2c3d4e5'
  /auth/reset-password:
    post:
      summary: Reset password
      tags:
        - Authentication
      description: Resets the user's password using the reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                resetToken:
                  type: string
                  example: 'a1b2c3d4e5'
                newPassword:
                  type: string
                  format: password
                  example: 'NewPassword123!'
                confirmNewPassword:
                  type: string
                  format: password
                  example: 'NewPassword123!'
              required:
                - email
                - resetToken
                - newPassword
                - confirmNewPassword
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password reset successfully
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - username
        - password
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        email:
          type: string
        gender:
          type: string
          enum: [Male, Female]
        dob:
          type: string
          format: date
        phone:
          type: string
      required:
        - username
        - password
        - email
        - dob
        - phone
        - gender
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the operation was successful
          example: true
        message:
          type: string
          description: Success message
        data:
          type: object
          description: Additional data returned by the operation
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...